requirements.txt
fastapi==0.121.3
httpx==0.28.1
starlette==0.50.0
python-dotenv>=1.0.0
requests==2.32.5
huggingface_hub==1.1.4
coverage==7.12.0
watchtower==3.4.0
boto3==1.40.76
botocore==1.40.76
mysql-connector-python==9.5.0
pytest==9.0.1
pytest-cov==7.0.0
pydantic==2.12.4
sqlmodel==0.0.27
jinja2==3.1.6
uvicorn==0.38.0
python-multipart==0.0.20
pymysql==1.1.2
docker==7.1.0
redis==7.1.0
GitPython==3.1.45
SQLAlchemy~=2.0.44
huggingface_hub~=1.1.4
pydantic_core~=2.41.5
typing_extensions~=4.15.0
dotenv~=0.9.9
cryptography~=46.0.3
huggingface_hub~=1.1.4
pylint==4.0.3
beautifulsoup4
